Got a basic implementation working
